Move vosk functionality to its own file
